convert test files
"validation.test.ts" and "error-validation.test.ts" is not yet converted.
