+++ packages/graph-schema-utils/test/validation/full.test.ts
@@ -1,6 +1,6 @@
 import { strict as assert } from "node:assert";
 import path from "path";
-import { validateSchema } from "../../src/index.ts";
+import { validateSchema } from "../../src/index";
 import { readFile } from "../fs.utils.js";
 import { describe, test } from "vitest";
 import { createRequire } from "module";
