+++ packages/graph-schema-utils/test/validation/new-props.test.ts
@@ -1,7 +1,6 @@
 import { strict as assert } from "node:assert";
-import { fileURLToPath } from "url";
 import path from "path";
-import { validateSchema } from "../../src/index.ts";
+import { validateSchema } from "../../src/index";
 import { readFile } from "../fs.utils.js";
 import { describe, test } from "vitest";
 import { createRequire } from "module";
@@ -11,13 +10,12 @@ const JSON_SCHEMA = JSON.stringify(
   require("@neo4j/graph-json-schema/json-schema.json")
 );
 
-// Happy path == schemas we expect to pass
-describe("Simple happy path tests", () => {
-  test("Single node test", () => {
-    const minimalGraphSchema = readFile(
-      path.resolve(__dirname, "./test-schemas/single-node.json")
+describe("Props", () => {
+  test("New pros", () => {
+    const dbSchema = readFile(
+      path.resolve(__dirname, "./test-schemas/new-properties-format.json")
     );
-    const valid = validateSchema(JSON_SCHEMA, minimalGraphSchema);
+    const valid = validateSchema(JSON_SCHEMA, dbSchema);
     assert.strictEqual(valid, true);
   });
 });
