+++ packages/graph-schema-utils/test/validation/root-schema.test.ts
@@ -1,6 +1,6 @@
 import { strict as assert } from "node:assert";
 import path from "path";
-import { validateSchema } from "../../src/index.ts";
+import { validateSchema } from "../../src/index";
 import { readFile } from "../fs.utils.js";
 import { describe, test } from "vitest";
 import { createRequire } from "module";
@@ -10,12 +10,14 @@ const JSON_SCHEMA = JSON.stringify(
   require("@neo4j/graph-json-schema/json-schema.json")
 );
 
-describe("Props", () => {
-  test("New pros", () => {
-    const dbSchema = readFile(
-      path.resolve(__dirname, "./test-schemas/new-properties-format.json")
+// Validate root schema, allowing root schema.
+describe("Validate that root schema is allowed ", () => {
+  test("Validates root schema", () => {
+    const allowRootSchema = readFile(
+      path.resolve(__dirname, "./test-schemas/allowing-root-schema-field.json")
     );
-    const valid = validateSchema(JSON_SCHEMA, dbSchema);
+
+    const valid = validateSchema(JSON_SCHEMA, allowRootSchema);
     assert.strictEqual(valid, true);
   });
 });
