+++ packages/graph-schema-utils/test/validation/simple.test.ts
@@ -1,6 +1,7 @@
 import { strict as assert } from "node:assert";
+import { fileURLToPath } from "url";
 import path from "path";
-import { validateSchema } from "../../src/index.ts";
+import { validateSchema } from "../../src/index";
 import { readFile } from "../fs.utils.js";
 import { describe, test } from "vitest";
 import { createRequire } from "module";
@@ -10,14 +11,13 @@ const JSON_SCHEMA = JSON.stringify(
   require("@neo4j/graph-json-schema/json-schema.json")
 );
 
-// Validate root schema, allowing root schema.
-describe("Validate that root schema is allowed ", () => {
-  test("Validates root schema", () => {
-    const allowRootSchema = readFile(
-      path.resolve(__dirname, "./test-schemas/allowing-root-schema-field.json")
+// Happy path == schemas we expect to pass
+describe("Simple happy path tests", () => {
+  test("Single node test", () => {
+    const minimalGraphSchema = readFile(
+      path.resolve(__dirname, "./test-schemas/single-node.json")
     );
-
-    const valid = validateSchema(JSON_SCHEMA, allowRootSchema);
+    const valid = validateSchema(JSON_SCHEMA, minimalGraphSchema);
     assert.strictEqual(valid, true);
   });
 });
